requirements.txt
pillow==10.2.0
pyinstaller==6.3.0
